+++ CHANGES.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - LoRA.
 - Transformer weights dtype.
 - User-requested max tokenizer length (overrides HF max model length).
+- GPU memory tracking for memory debugging.
 
 ### Added